Add missing include of value.h
With MSVC we have following problem:

in cmd_memory.c value.h is included AFTER cmd_peek declaration. So compiler
thinks that TypedValue is a large anonymous structure and expects pointer to
recipient to be on the stack.

interpreter.c includes value.h before cmd_peek declaration and compiler
sees it's a small structure and optimizes by returning it in registers.

When both collides, cmd_peek looks for core argument in wrong place leads to
crash by accessing a bad pointer.